requirements.txt
torch>=2.0.0
transformers>=4.30.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
mysql-connector-python>=8.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
accelerate>=0.26.0
